+++ src/plugins/core/metaData/refMetaDataHandler.ts
@@ -1,7 +1,6 @@
 import { log } from '../../../io';
 import * as models from '../../../models';
 import * as utils from '../../../utils';
-import { ImportProcessorContext } from './importMetaDataHandler';
 
 export function refMetaDataHandler(type: string, name: string | undefined, context: models.ParserContext): boolean {
   if (['ref', 'forceRef'].indexOf(type) >= 0 && name) {
@@ -27,64 +26,22 @@ class RefMetaAction {
 
   constructor(private readonly data: RefMetaHttpRegionData) {}
 
-  async process(context: ImportProcessorContext): Promise<boolean> {
+  async process(context: models.ProcessorContext): Promise<boolean> {
+    let result = true;
     utils.report(context, `load reference ${this.data.name}`);
-    let reference = this.findHttpRegionInContext(context.httpFile);
-    if (!reference) {
-      reference = this.findHttpRegionInImportedContext(context);
-    }
-
+    const reference = utils.findHttpRegionInContext(this.data.name, context);
     if (reference) {
-      utils.registerRegionDependent(
-        context,
-        reference.httpFile,
-        reference.httpRegion,
-        context.httpFile,
-        context.httpRegion
-      );
       const envKey = utils.toEnvironmentKey(context.httpFile.activeEnvironment);
-      log.trace('import variables', reference.httpRegion.variablesPerEnv[envKey]);
-      utils.setVariableInContext(reference.httpRegion.variablesPerEnv[envKey], context);
+      utils.setVariableInContext(reference.variablesPerEnv[envKey], context);
+      log.trace('import variables', reference.variablesPerEnv[envKey]);
       if (this.data.force || utils.isUndefined(context.variables[this.data.name])) {
-        const refContext = {
-          ...context,
-          httpFile: reference.httpFile,
-          httpRegion: reference.httpRegion,
-          isMainContext: false,
-        };
-        const result = await utils.processHttpRegionActions(refContext);
+        result = await reference.execute(context);
         if (result) {
-          utils.setVariableInContext(refContext.httpRegion.variablesPerEnv[envKey], context);
-        }
-        return result;
-      }
-      return true;
-    }
-    return true;
-  }
-
-  private findHttpRegionInContext(httpFile: models.HttpFile) {
-    const httpRegion = httpFile.httpRegions.find(
-      refHttpRegion => refHttpRegion.metaData.name === this.data.name && !refHttpRegion.metaData.disabled
-    );
-    if (httpRegion) {
-      return {
-        httpRegion,
-        httpFile,
-      };
-    }
-    return undefined;
-  }
-
-  private findHttpRegionInImportedContext(context: ImportProcessorContext) {
-    if (context.options.httpFiles) {
-      for (const { ref } of context.options.httpFiles.filter(obj => obj.base === context.httpFile)) {
-        const reference = this.findHttpRegionInContext(ref);
-        if (reference) {
-          return reference;
+          // ref to ref variable export
+          utils.setVariableInContext(reference.variablesPerEnv[envKey], context);
         }
       }
     }
-    return undefined;
+    return result;
   }
 }