+++ src/store/index.ts
@@ -1,3 +1,4 @@
 export * from './httpFileStore';
 export * from './pluginStore';
 export * from './userSessionStore';
+export * from './getEnvironmentConfig';