+++ src/utils/index.ts
@@ -14,6 +14,7 @@ export * from './mimeTypeUtils';
 export * from './modelUtils';
 export * from './parseRequestClientFactory';
 export * from './parserUtils';
+export * from './processorContextUtils';
 export * from './promiseUtils';
 export * from './randomUtils';
 export * from './repeatUtils';